implemented application auditing

+++ book-social-network/src/main/java/com/booknetwork/api/common/BaseEntity.java
@@ -42,9 +42,9 @@ public class BaseEntity {
 
     @CreatedBy
     @Column(nullable = false, updatable = false)
-    private Integer createdBy;
+    private Long createdBy;
 
     @LastModifiedBy
     @Column(insertable = false)
-    private Integer lastModifiedBy;
+    private Long lastModifiedBy;
 }
